Calculating method hashes
